Did some alterations on CreateVMAction.java,ProcessLifecycle.java and CloudProcess.java so when a CloudProcess is added in the database it sets the following values of CloudProcess: Account, Epoch and costPerHours. Also made some tests in the CreateVMActionTest.java so it's possible to run some tests of those new additions.

+++ n3phele/src/n3phele/service/model/CloudProcess.java
@@ -64,7 +64,7 @@ public class CloudProcess extends Entity {
 	@Parent Key<CloudProcess> root;
 	@Index(IfTrue.class) boolean topLevel = false;
 	@Index(IfTrue.class) boolean haveCost = false;
-	@Index(IfNotZero.class)protected float costPerHour = 0;
+	@Index(IfNotZero.class)protected double costPerHour = 0;
 	protected Date epoch = null;
 	protected String account = null;
 
@@ -128,14 +128,14 @@ public class CloudProcess extends Entity {
 	/**
 	 * @return the cost per hour
 	 */
-	public float getCostPerHour(){
+	public double getCostPerHour(){
 		return costPerHour;
 	}
 	
 	/**
 	 * @param costPerHour the cost per hour to set
 	 */
-	public void setCostPerHour(float costPerHour){
+	public void setCostPerHour(double costPerHour){
 		this.costPerHour = costPerHour;
 		this.haveCost = (costPerHour > 0);
 	}
